options_agent/requirements.txt
# MCP (for brokerage connection)
mcp>=1.0.0

# LLM (optional — for strategy confirmation via Gemini)
google-genai>=1.0.0

# Brokerage
robin_stocks>=3.0.0
pyotp>=2.9.0

# Market Data & Analysis
yfinance>=0.2.40
alpaca-py>=0.30.0
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0
scipy>=1.12.0

# Data Models & Config
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Logging
structlog>=24.4.0

# Web / Dashboard
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.38.0
streamlit-autorefresh>=1.0.1
plotly>=5.22.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

